+++ src/common/web/component/MetaInformation.ts
@@ -3,8 +3,6 @@ import metaData from "/config/meta-information.json";
 
 import { SemVer } from "semver";
 
-import { Configuration } from "../utils/config/Configuration";
-
 type ComponentInformationType = {
     name: string;
     directory: string;
@@ -16,20 +14,13 @@ type MetaInformationType = {
         title: string;
         version: string;
     };
-    components: Record<string, ComponentInformationType>;
 };
 
 /**
- * Accesses meta information about the entire project and its various component stored in a *JSON* file.
+ * Accesses meta information about the entire project stored in a *JSON* file.
  */
 export class MetaInformation {
-    private readonly _config: Configuration;
-
     private readonly _data = metaData as MetaInformationType;
-
-    public constructor(config: Configuration) {
-        this._config = config;
-    }
 
     /**
      * The project title.
@@ -43,26 +34,5 @@ export class MetaInformation {
      */
     public get version(): SemVer {
         return new SemVer(this._data.global.version);
-    }
-
-    /**
-     * Retrieves the meta information stored for a specific component.
-     *
-     * This meta information includes the ``name`` of the component, as well as its ``directory`` within the code structure (rooted at ``/src``).
-     *
-     * @param comp - The name of the component.
-     *
-     * @returns - A dictionary containing the meta information.
-     */
-    public getComponent(comp: string): ComponentInformationType {
-        if (comp in this._data.components) {
-            return this._data.components[comp]!;
-        }
-
-        return {
-            name: "<invalid>",
-            directory: "",
-            tech: ""
-        };
     }
 }