+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Improvements
 
+- Meta and connector information can be set via environment variables
 - Configurable support email address
 
 ### Fixes